+++ CHANGELOG.md
@@ -6,7 +6,11 @@
 - It is possible to configure a disaster-recovery site and display its status
   ([rhbz#1676431])
 
+### Fixed
+- Error messages in cases when cluster is not set up ([rhbz#1743731])
+
 [rhbz#1676431]: https://bugzilla.redhat.com/show_bug.cgi?id=1676431
+[rhbz#1743731]: https://bugzilla.redhat.com/show_bug.cgi?id=1743731
 
 
 ## [0.10.4] - 2019-11-28